cutscene & skeleton reorg
