+++ fast64_internal/oot/cutscene/constants.py
@@ -1,0 +1,31 @@
+ootEnumCSTextboxTypeEntryC = {
+    "Text": "CS_TEXT_DISPLAY_TEXTBOX",
+    "None": "CS_TEXT_NONE",
+    "LearnSong": "CS_TEXT_LEARN_SONG",
+}
+
+ootEnumCSListTypeListC = {
+    "Textbox": "CS_TEXT_LIST",
+    "FX": "CS_SCENE_TRANS_FX",
+    "Lighting": "CS_LIGHTING_LIST",
+    "Time": "CS_TIME_LIST",
+    "PlayBGM": "CS_PLAY_BGM_LIST",
+    "StopBGM": "CS_STOP_BGM_LIST",
+    "FadeBGM": "CS_FADE_BGM_LIST",
+    "Misc": "CS_MISC_LIST",
+    "0x09": "CS_CMD_09_LIST",
+    "Unk": "CS_UNK_DATA_LIST",
+}
+
+ootEnumCSListTypeEntryC = {
+    "Textbox": None,  # special case
+    "FX": None,  # no list entries
+    "Lighting": "CS_LIGHTING",
+    "Time": "CS_TIME",
+    "PlayBGM": "CS_PLAY_BGM",
+    "StopBGM": "CS_STOP_BGM",
+    "FadeBGM": "CS_FADE_BGM",
+    "Misc": "CS_MISC",
+    "0x09": "CS_CMD_09",
+    "Unk": "CS_UNK_DATA",
+}
