+++ fast64_internal/oot/cutscene/exporter/__init__.py
@@ -1,0 +1,13 @@
+from .cutscene import convertCutsceneObject, readCutsceneData
+
+from .classes import (
+    OOTCSList,
+    OOTCSTextbox,
+    OOTCSLighting,
+    OOTCSTime,
+    OOTCSBGM,
+    OOTCSMisc,
+    OOTCS0x09,
+    OOTCSUnk,
+    OOTCutscene,
+)
