+++ fast64_internal/oot/scene/exporter/to_c/__init__.py
@@ -3,3 +3,4 @@ from .scene_table_c import modifySceneTable, getDrawConfig
 from .spec import editSpecFile
 from .scene_folder import modifySceneFiles, deleteSceneFiles
 from .scene_bootup import setBootupScene, clearBootupScene
+from .scene_cutscene import ootCutsceneDataToC
