+++ fast64_internal/oot/skeleton/exporter/__init__.py
@@ -1,0 +1,1 @@
+from .skeleton import ootConvertArmatureToSkeletonWithoutMesh, ootConvertArmatureToC
